Revert "mem: reset SOFT_DIRTY bits after dumping of all vmas in the process"
This reverts commit 101e3ba49d0801511a50241a3e2f6ce17f7641a4

CRIU uses proc pagemaps for 2 actions:
- Save ashmem pagemap flags in __parasite_dump_pages_seized()
into its own bitmaps.
- Process private anon mem pagemaps in __parasite_dump_pages_seized().
After these 2 actions pagemaps are not needed and we can call
task_reset_dirty_track().

This analysis shows that we don't need movement of
task_reset_dirty_track() call introduced in commit
101e3ba49d0801511a50241a3e2f6ce17f7641a4.
Revert this commit because it is suspected to be the cause of bug #203.

Signed-off-by: Eugene Batalov <eabatalov89@gmail.com>
Acked-by:  Andrei Vagin <avagin@virtuozzo.com>
Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>